# go/mysql: performance optimizations in protocol encoding