@dbussink 2024-07-05
@mattrobenolt Is clearing this way faster than using the following?

```go
func writeZeroes(data []byte, pos int, len int) int {
	data = data[pos : pos+len]
	for i := range data {
		data[i] = 0
	}
	return pos + len
}
```

The Go compiler recognizes that pattern and optimizes it into a `CALL runtime.memclrNoHeapPointers(SB)` essentially. Is that just as fast or even faster? See also https://github.com/golang/go/issues/5373 for that pattern.