@mattrobenolt 2024-07-05
@dbussink lol so

```go
package main

import "testing"

func writeZeroesVitessMain(data []byte, pos, len int) int {
	for i := 0; i < len; i++ {
		data[pos+i] = 0
	}
	return pos + len
}

func writeZeroesSpecialized23(data []byte, pos int) int {
	data = data[pos:]

	_ = data[22]
	data[0] = 0
	data[1] = 0
	data[2] = 0
	data[3] = 0
	data[4] = 0
	data[5] = 0
	data[6] = 0
	data[7] = 0
	data[8] = 0
	data[9] = 0
	data[10] = 0
	data[11] = 0
	data[12] = 0
	data[13] = 0
	data[14] = 0
	data[15] = 0
	data[16] = 0
	data[17] = 0
	data[18] = 0
	data[19] = 0
	data[20] = 0
	data[21] = 0
	data[22] = 0

	return pos + 23
}

func writeZeroesSpecialized10(data []byte, pos int) int {
	data = data[pos:]

	_ = data[9]
	data[0] = 0
	data[1] = 0
	data[2] = 0
	data[3] = 0
	data[4] = 0
	data[5] = 0
	data[6] = 0
	data[7] = 0
	data[8] = 0
	data[9] = 0

	return pos + 10
}

func writeZeroesMemclr(data []byte, pos, len int) int {
	end := pos + len
	data = data[pos:end]

	for i := range data {
		data[i] = 0
	}

	return end
}

func BenchmarkZeroes(b *testing.B) {
	buf := make([]byte, 128)

	b.Run("vitess-main/23-byte", func(b *testing.B) {
		for range b.N {
			_ = writeZeroesVitessMain(buf, 16, 23)
		}
	})

	b.Run("vitess-main/10-byte", func(b *testing.B) {
		for range b.N {
			_ = writeZeroesVitessMain(buf, 16, 10)
		}
	})

	b.Run("specialized/23-byte", func(b *testing.B) {
		for range b.N {
			_ = writeZeroesSpecialized23(buf, 16)
		}
	})

	b.Run("specialized/10-byte", func(b *testing.B) {
		for range b.N {
			_ = writeZeroesSpecialized10(buf, 16)
		}
	})

	b.Run("memclr/23-byte", func(b *testing.B) {
		for range b.N {
			_ = writeZeroesMemclr(buf, 16, 23)
		}
	})

	b.Run("memclr/10-byte", func(b *testing.B) {
		for range b.N {
			_ = writeZeroesMemclr(buf, 16, 10)
		}
	})
}
```

I put them all side by side and this is wild.

```
$ go test -v . -bench=.
goos: darwin
goarch: arm64
pkg: x
BenchmarkZeroes
BenchmarkZeroes/vitess-main/23-byte
BenchmarkZeroes/vitess-main/23-byte-10          138833698                8.634 ns/op
BenchmarkZeroes/vitess-main/10-byte
BenchmarkZeroes/vitess-main/10-byte-10          268413115                4.471 ns/op
BenchmarkZeroes/specialized/23-byte
BenchmarkZeroes/specialized/23-byte-10          570586390                2.125 ns/op
BenchmarkZeroes/specialized/10-byte
BenchmarkZeroes/specialized/10-byte-10          1000000000               0.6531 ns/op
BenchmarkZeroes/memclr/23-byte
BenchmarkZeroes/memclr/23-byte-10               1000000000               0.3263 ns/op
BenchmarkZeroes/memclr/10-byte
BenchmarkZeroes/memclr/10-byte-10               1000000000               0.3286 ns/op
PASS
ok      x       7.020s
```

That memclr optimization is really good.

Going to swap that out, which will let me get rid of the specialized versions. I didn't like that anyways.