@mattrobenolt 2024-07-05
@dbussink updated to the memclr optimization as well as benchmarks in the PR description.